requirements.txt
numpy==1.19.1
opencv_python==4.4.0.42
tensorflow==2.2.0
jsonify==0.5
imageio==2.9.0
Flask==1.1.2
Pillow==7.2.0
scikit_learn==0.23.2
gunicorn
